refactor DirectFsAccess
