+++ src/databricks/labs/ucx/source_code/graph.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import abc
 import itertools
-import json
 import logging
 from dataclasses import dataclass
 from pathlib import Path
@@ -13,6 +12,7 @@ from astroid import (  # type: ignore
     NodeNG,
 )
 from databricks.labs.ucx.source_code.base import Advisory, CurrentSessionState, is_a_notebook
+from databricks.labs.ucx.source_code.directfs_access import LineageAtom
 from databricks.labs.ucx.source_code.python.python_ast import Tree
 from databricks.labs.ucx.source_code.path_lookup import PathLookup
 
@@ -605,23 +605,6 @@ class InheritedContext:
         return InheritedContext(tree, self.found)
 
 
-@dataclass
-class LineageAtom:
-
-    @staticmethod
-    def atoms_to_json_string(atoms: list[LineageAtom]):
-        json_lists = list(lineage.as_objects() for lineage in atoms)
-        json_obj = list(itertools.chain(*json_lists))
-        return json.dumps(json_obj)
-
-    object_type: str
-    object_id: str
-    other: dict[str, str] | None = None
-
-    def as_objects(self) -> list[dict[str, str]]:
-        return [{"object_type": self.object_type, "object_id": self.object_id, **(self.other or {})}]
-
-
 T = TypeVar("T")
 
 
@@ -668,5 +651,5 @@ class DependencyGraphWalker(abc.ABC, Generic[T]):
 
     @property
     def lineage(self) -> list[LineageAtom]:
-        lineages = [dependency.lineage for dependency in self._lineage]
-        return list(itertools.chain(*lineages))
+        lists: list[list[LineageAtom]] = [dependency.lineage for dependency in self._lineage]
+        return list(itertools.chain(*lists))
